CteateLoginAndRegistr

+++ ProniaOnion202/SRC/Infrastructure/ProniaOnion202.Persistance/Implementations/Services/ProductService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Conventions;
 using ProniaOnion202.Applicatin.Abstractions.Repositories;
 using ProniaOnion202.Applicatin.Abstractions.Services;
 using ProniaOnion202.Applicatin.DTOs.Products;
@@ -12,30 +13,30 @@ using System.Threading.Tasks;
 
 namespace ProniaOnion202.Persistance.Implementations.Services
 {
-    internal class ProductService : IProductService
+    public class ProductService : IProductService
     {
         private readonly IProductRepository _repo;
         private readonly ICategoryRepository _catrep;
         private readonly IColorRepository _colrepo;
         private readonly IMapper _map;
 
-        public ProductService(IProductRepository repo,ICategoryRepository catrep ,IColorRepository colrepo,IMapper map)
+        public ProductService(IProductRepository repo, ICategoryRepository catrep, IColorRepository colrepo, IMapper map)
         {
             _repo = repo;
             _catrep = catrep;
             _colrepo = colrepo;
             _map = map;
         }
-        public async Task <IEnumerable<ProductItemDto>> GetAllAsync(int page ,int take)
+        public async Task<IEnumerable<ProductItemDto>> GetAllAsync(int page, int take)
         {
-           
+
             return _map.Map<IEnumerable<ProductItemDto>>(await _repo.GetAllWhere(skip: (page - 1) * take, take: take).ToListAsync());
         }
         public async Task<ProductGetDto> GetByIdAsync(int id)
         {
-            Product product=await _repo.GetByIdAsync(id,includes:nameof(Product.Category));
+            Product product = await _repo.GetByIdAsync(id, includes: nameof(Product.Category));
             if (product is null) throw new Exception("Not Found");
-           
+
             return _map.Map<ProductGetDto>(product);
         }
         public async Task CreateAsync(ProductCreateDto dto)
@@ -43,9 +44,9 @@ namespace ProniaOnion202.Persistance.Implementations.Services
 
             if (await _repo.IsExistAsync(p => p.Name == dto.Name)) throw new Exception("Same");
             if (!await _catrep.IsExistAsync(c => c.Id == dto.CategoryId)) throw new Exception("Dont");
-            Product product=_map.Map<Product>(dto);
-            product.ProductColors=new List<ProductColor>();
-            if(dto.Colorids is not null)
+            Product product = _map.Map<Product>(dto);
+            product.ProductColors = new List<ProductColor>();
+            if (dto.Colorids is not null)
             {
                 foreach (var item in dto.Colorids)
                 {
@@ -55,19 +56,19 @@ namespace ProniaOnion202.Persistance.Implementations.Services
             }
             await _repo.AddAsync(product);
             await _repo.SaveChangesAsync();
-            
+
         }
-        public async Task UpdateAsync(int id,ProductPutDto dto)
+        public async Task UpdateAsync(int id, ProductPutDto dto)
         {
-            Product exist = await _repo.GetByIdAsync(id, true,includes:nameof(Product.ProductColors));
-            if (exist.Name!=dto.Name)
-                 if (await _repo.IsExistAsync(p => p.Name == dto.Name)) throw new Exception("Same");
+            Product exist = await _repo.GetByIdAsync(id, true, includes: nameof(Product.ProductColors));
+            if (exist.Name != dto.Name)
+                if (await _repo.IsExistAsync(p => p.Name == dto.Name)) throw new Exception("Same");
 
             if (dto.CategoryId != exist.CategoryId)
                 if (!await _catrep.IsExistAsync(c => c.Id == dto.CategoryId)) throw new Exception("Dont");
-            exist =_map.Map(dto,exist);
-           
-            if(dto.Colorids is not null)
+            exist = _map.Map(dto, exist);
+
+            if (dto.Colorids is not null)
             {
                 foreach (var item in dto.Colorids)
                 {
@@ -82,18 +83,47 @@ namespace ProniaOnion202.Persistance.Implementations.Services
             }
             else
             {
-                exist.ProductColors=new List<ProductColor>();
+                exist.ProductColors = new List<ProductColor>();
             }
-           
+
             _repo.Update(exist);
             await _repo.SaveChangesAsync();
         }
         public async Task DeleteAsync(int id)
         {
             Product exist = await _repo.GetByIdAsync(id);
-
-
+            if (exist == null) throw new Exception("Not Found");
             _repo.Delete(exist);
+            await _repo.SaveChangesAsync();
+        }
+        public async Task SoftDeleteAsync(int id)
+        {
+            Product exist = await _repo.GetByIdAsync(id, true, includes: nameof(Product.ProductColors));
+            if (exist == null) throw new Exception("Not Found");
+            ICollection<ProductColor> pcolors = exist.ProductColors.Where(p => p.ProductId == id).ToList();
+            if (pcolors!=null) 
+            { 
+                foreach (var color in pcolors)
+                {
+                    if(!color.IsDeleted)color.IsDeleted = true;
+                }
+            }
+            _repo.SoftDelete(exist);
+            await _repo.SaveChangesAsync();
+        }
+        public async Task ReverseAsync(int id)
+        {
+            Product exist = await _repo.GetByIdAsync(id,true, ignorQuery:true,includes: nameof(Product.ProductColors));
+            if (exist == null) throw new Exception("Not Found");
+            ICollection<ProductColor> pcolors = exist.ProductColors.Where(p => p.ProductId == id).ToList();
+            if (pcolors != null)
+            {
+                foreach (var color in pcolors)
+                {
+                    if (!color.IsDeleted) color.IsDeleted = false;
+                }
+            }
+            _repo.ReverseDelete(exist);
             await _repo.SaveChangesAsync();
         }
     }
